Extract presto-testng-services module
