+++ presto-main/pom.xml
@@ -355,6 +355,12 @@
 
         <dependency>
             <groupId>io.prestosql</groupId>
+            <artifactId>presto-testng-services</artifactId>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>io.prestosql</groupId>
             <artifactId>presto-tpch</artifactId>
             <scope>test</scope>
         </dependency>
