--- presto-main/src/test/resources/META-INF/services/org.testng.ITestNGListener
@@ -1,3 +1,0 @@
-io.prestosql.tests.ReportUnannotatedMethods
-io.prestosql.tests.ReportIllNamedTest
-io.prestosql.tests.LogTestDurationListener
