+++ presto-testing/pom.xml
@@ -139,6 +139,12 @@
 
         <!-- used by tests but also needed transitively -->
         <dependency>
+            <groupId>io.prestosql</groupId>
+            <artifactId>presto-testng-services</artifactId>
+            <scope>runtime</scope>
+        </dependency>
+
+        <dependency>
             <groupId>io.airlift</groupId>
             <artifactId>log-manager</artifactId>
             <scope>runtime</scope>
