+++ presto-testng-services/src/main/java/io/prestosql/testng/services/LogTestDurationListener.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.tests;
+package io.prestosql.testng.services;
 
 import com.google.common.collect.ImmutableMap;
 import io.airlift.log.Logger;
