+++ presto-testng-services/src/main/java/io/prestosql/testng/services/ReportIllNamedTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.tests;
+package io.prestosql.testng.services;
 
 import org.testng.IClassListener;
 import org.testng.ITestClass;
